port Modal to ES6 as Trends to be able to put bgBounds on component instance to avoid excess rerenders due to new object each render
