+++ app/components/chart/index.js
@@ -17,7 +17,6 @@ module.exports = {
   basics: require('./basics'),
   daily: require('./daily'),
   weekly: require('./weekly'),
-  modal: require('./modal'),
   settings: require('./settings'),
   header: require('./header'),
   footer: require('./footer')
