+++ app/components/chart/trendssubnav.less
