+++ app/pages/patientdata/patientdata.js
@@ -33,7 +33,7 @@ import { URL_UPLOADER_CHROME_STORE, URL_BLIP_NOTES_APP_STORE } from '../../core/
 import { header as Header } from '../../components/chart';
 import { basics as Basics } from '../../components/chart';
 import { daily as Daily } from '../../components/chart';
-import { modal as Modal } from '../../components/chart';
+import Trends from '../../components/chart/trends';
 import { weekly as Weekly } from '../../components/chart';
 import { settings as Settings } from '../../components/chart';
 
@@ -270,7 +270,7 @@ export let PatientData = React.createClass({
       case 'modal':
 
         return (
-          <Modal
+          <Trends
             bgPrefs={this.state.bgPrefs}
             chartPrefs={this.state.chartPrefs}
             currentPatientInViewId={this.props.currentPatientInViewId}
