+++ app/style.less
@@ -30,7 +30,7 @@
 @import "components/modaloverlay/modaloverlay.less";
 @import "components/datepicker/datepicker.less";
 @import "components/waitlist/waitlist.less";
-@import "components/chart/modalsubnav.less";
+@import "components/chart/trendssubnav.less";
 @import "components/browserwarning/browserwarning.less";
 @import "components/uploaderbutton/uploaderbutton.less";
